Add move-out settlement feature with APIs and services
Introduced a comprehensive move-out settlement feature:
- Added `SettlementController` with endpoints for settlement management.
- Registered `ISettlementService` and `ISettlementRepository` in DI.
- Implemented `SettlementService` and `SettlementRepository`.
- Added new domain entities: `MoveOutSettlementInfo`, `SettlementPendingActionInfo`.
- Created request/response models for settlement operations.
- Updated `DBConstants` with stored procedures for settlements.
- Enhanced existing entities (`FilterInfo`, `RentPaymentInfo`, `UtilityChargeInfo`).
- Integrated stored procedures for settlement-related operations.
- Commented out unused `CreateSettlement` endpoint in `SettlementController`.

+++ FEMOS.Rentora.Domain/Entities/UtilityChargeInfo.cs
@@ -15,7 +15,7 @@ namespace FEMOS.Rentora.Domain.Entities
         public long TenantAssignmentId { get; set; }
         public string? TenantName { get; set; }
         public string? UnitNumber { get; set; }
-        public short UtilityTypeId { get; set; }
+        public int UtilityTypeId { get; set; }
         public string? UtilityType { get; set; }
         public bool IsMeterBased { get; set; }
         public DateTime ChargeDate { get; set; }
